pyproject.toml
[project]
name = "agent-orchestrator"
version = "0.1.0"
description = "Multi-model agent orchestration service via OpenRouter"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
agent-orchestrator = "src.main:main"
